Add `useBitcoinAddresses` hook
