+++ apps/extension/src/hooks/bitcoin/use-utxos.ts
@@ -1,8 +1,5 @@
-import { useMemo } from "react";
 import { useStore } from "../../stores";
-import { payments } from "bitcoinjs-lib";
-import { toXOnly } from "@keplr-wallet/crypto";
-import { useBitcoinNetworkConfig } from "./use-bitcoin-network-config";
+import { useBitcoinAddresses } from "./use-bitcoin-network-config";
 
 export const useUTXOs = (chainId: string, address: string) => {
   const { chainStore, bitcoinQueriesStore } = useStore();
@@ -21,43 +18,12 @@ export const useUTXOs = (chainId: string, address: string) => {
 };
 
 export const useNativeSegwitUTXOs = (chainId: string) => {
-  const { accountStore } = useStore();
-  const { networkConfig } = useBitcoinNetworkConfig(chainId);
-
-  const nativeSegwitAddress = useMemo(() => {
-    const pubkey = accountStore.getAccount(chainId).pubKey;
-    if (!pubkey) {
-      return undefined;
-    }
-
-    const { address } = payments.p2wpkh({
-      pubkey: Buffer.from(pubkey),
-      network: networkConfig,
-    });
-
-    return address;
-  }, [accountStore, chainId, networkConfig]);
+  const { nativeSegwitAddress } = useBitcoinAddresses(chainId);
 
   return useUTXOs(chainId, nativeSegwitAddress ?? "");
 };
 
 export const useTaprootUTXOs = (chainId: string) => {
-  const { accountStore } = useStore();
-  const { networkConfig } = useBitcoinNetworkConfig(chainId);
-
-  const taprootAddress = useMemo(() => {
-    const pubkey = accountStore.getAccount(chainId).pubKey;
-    if (!pubkey) {
-      return undefined;
-    }
-
-    const { address } = payments.p2tr({
-      internalPubkey: toXOnly(Buffer.from(pubkey)),
-      network: networkConfig,
-    });
-
-    return address;
-  }, [accountStore, chainId, networkConfig]);
-
+  const { taprootAddress } = useBitcoinAddresses(chainId);
   return useUTXOs(chainId, taprootAddress ?? "");
 };
